Speed up data_dir load by making ConfigurationPasswordEncryptionHelper faster
ConfigurationPasswordEncryptionHelper.decode(String)
is called by XstreamPersister -> StoreInfoConverter
for each DataStoreInfo read from the data directory at
start up time.

This patch makes it use a ConcurrentMap internally
instead of a HashMap, and removes the sychronization
on its CACHE member variable. Additionally, and most
importantly, caches the encryptable store parameter
names by StoreInfo.getType() too, avoiding the fallback
to full DataAccessFactory parameter lookup, which
implies a rather expensive call to getCatalog() \
.getResourcePool().getDataStoreFactory((DataStoreInfo) info)

As a result, loading time when there are lots of
DataStoreInfos is reduced to about 1/3 of the time
it takes without this patch.